Decided against using enums and the movement class
Created an array of directions based on the left or right turn progression. Decided against using enums as it added an unnecessary layer of complexity. Removed the movement class and decided to do all work in the program file because was not sure how to pass through the coordinate and direction details without getting really messy. So now can use just the local variables and update them when necessary. 
Potentially not the best way to do this. Will update if required.