Updated Place switch option
There is now copied broken code so will need to rectify that. Potential problem if the program refreshes every time it runs, will the local variables be lost and changed every time. There is some confusion as to how to get the new prompting, does it need a loop to continually be running the switch statement. 
Also Left and Right would not take additional fields of input so will this crash the system??

Will investigate further