Fixed report functionality
Passed the single instance of bot through the handleCommands() function meaning that we only ever create one instance of bot. 
Successfully passed information and used hasbeenplaced because we are now referencing the one bot.
Unsure if this is the most efficient way of approaching this. Review after other functionalities implemented